test(unit): merge Cassandra cluster tests into unit tree
unit_tests/unit/test_cluster_cassandra.py: add JVM heap, version, no-op, and node-list coverage migrated from the misplaced root test module. Replace constructor patching with object.__new__ helper setup. Remove unit_tests/test_cluster_cassandra.py after merging its coverage.
